--- Editor/ZedPreferences.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 96c771f2489c84be28d6c3e4ea52e2d5
