--- Editor/ZedProcess.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 07a81d15fbf03419cb72889bbfbde513
