--- Editor/ZedSettings.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: f98a38ae53cba4c3ba4df26bbab3479f
